% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPathwayArchive.R
\name{downloadPathwayArchive}
\alias{downloadPathwayArchive}
\title{Download Pathway Archive}
\usage{
downloadPathwayArchive(date = "current", organism = NULL,
  format = c("gpml", "gmt", "svg"))
}
\arguments{
\item{date}{(optional) The timestamp for a monthly release (e.g., 20171010) 
or "current" (default) for latest release.}

\item{organism}{(optional) A particular species. See \link{listOrganisms}.}

\item{format}{(optional) Either gpml (default), gmt or svg.}
}
\value{
Downloaded file or tab in default browser
}
\description{
Access the monthly archives of pathway content from WikiPathways.
}
\details{
If you specify an archive date, organism and format, then the archive file
will be downloaded. Otherwise, the archive will be opened in a tab in your
default browser.
}
\examples{
\donttest{
downloadPathwayArchive()
downloadPathwayArchive(format="gmt")
downloadPathwayArchive(date="20171010", format="svg")
downloadPathwayArchive(date="20171010", organism="Mus musculus", format="svg")
}
}
