% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwaysByLiterature.R
\name{findPathwayIdsByLiterature}
\alias{findPathwayIdsByLiterature}
\title{Find Pathway WPIDs By Literature}
\usage{
findPathwayIdsByLiterature(query)
}
\arguments{
\item{query}{The \code{character} string to search for, e.g., a PMID, title keyword or author name.}
}
\value{
A \code{list} of WPIDs
}
\description{
Retrieve list of pathway WPIDs containing the query citation.
}
\examples{
{
findPathwayIdsByLiterature('19649250')
findPathwayIdsByLiterature('smith')
findPathwayIdsByLiterature('cancer')
}
}
