% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwaysByOrcid.R
\name{findPathwayIDsByOrcid}
\alias{findPathwayIDsByOrcid}
\title{Find Pathway WPIDs By ORCID}
\usage{
findPathwayIDsByOrcid(query = NULL)
}
\arguments{
\item{query}{The \code{character} ORCID to search for.}
}
\value{
A \code{list} of WPIDs
}
\description{
Retrieve list of pathway WPIDs containing the query ORCID
}
\examples{
{
findPathwayIDsByOrcid(' 0000-0001-9773-4008')
}
}
\seealso{
findPathwaysByOrcid
}
