% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwayByText.R
\name{findPathwayIdsByText}
\alias{findPathwayIdsByText}
\title{Find Pathway WPIDs By Text}
\usage{
findPathwayIdsByText(query)
}
\arguments{
\item{query}{The string to search for, e.g., "cancer"
e.g., En (Ensembl), L (Entrez), Ch (HMDB), etc.
See column two of https://github.com/bridgedb/BridgeDb/blob/master/org.bridgedb.bio/resources/org/bridgedb/bio/datasources.txt.}
}
\value{
List of WPIDs
}
\description{
Retrieve list of pathway WPIDs containing the query text.
}
\examples{
\donttest{
findPathwayIdsByText('cancer')
}
}
