% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwaysByOrcid.R
\name{findPathwayNamesByOrcid}
\alias{findPathwayNamesByOrcid}
\title{Find Pathway Names By ORCID}
\usage{
findPathwayNamesByOrcid(query = NULL)
}
\arguments{
\item{query}{The \code{character} ORCID to search for.}
}
\value{
A \code{list} of lists
}
\description{
Retrieve list of pathway names containing the query ORCID
}
\examples{
{
findPathwayNamesByOrcid(' 0000-0001-9773-4008')
}
}
\seealso{
findPathwaysByOrcid
}
