% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwaysByText.R
\name{findPathwayNamesByText}
\alias{findPathwayNamesByText}
\title{Find Pathway Names By Text}
\usage{
findPathwayNamesByText(query = NULL, field = NULL)
}
\arguments{
\item{query}{A \code{character} string to search for, e.g., "cancer"}

\item{field}{Optional \code{character} string to restrict search to a single
field, e.g., title, description, wpid, species, revision, authors,
datanodes, annotations, or citedIn.}
}
\value{
A \code{list} of pathway names
}
\description{
Retrieve list of pathway names containing the query text.
}
\details{
Note: there will be multiple listings of the same pathway name if 
copies exist for multiple species.
}
\examples{
{
findPathwayNamesByText('cancer')
}
}
\seealso{
findPathwaysByText
}
