% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwaysByLiterature.R
\name{findPathwaysByLiterature}
\alias{findPathwaysByLiterature}
\title{Find Pathways By Literature}
\usage{
findPathwaysByLiterature(query)
}
\arguments{
\item{query}{The \code{character} string to search for, e.g., a PMID, title keyword or author name.}
}
\value{
A \code{list} of lists
}
\description{
Retrieve a list of pathways containing the query citation.
}
\examples{
{
findPathwaysByLiterature('19649250')
findPathwaysByLiterature('smith')
findPathwaysByLiterature('cancer')
}
}
