% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwaysByOrcid.R
\name{findPathwaysByOrcid}
\alias{findPathwaysByOrcid}
\title{Find Pathways By ORCID}
\usage{
findPathwaysByOrcid(query = NULL)
}
\arguments{
\item{query}{The \code{character} ORCID to search for.}
}
\value{
A \code{dataframe} of pathway attributes including the matching
ORCIDs
}
\description{
Retrieve pathways containing the query ORCID
}
\examples{
{
findPathwaysByOrcid(' 0000-0001-9773-4008')
}
}
