% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwaysByOrcid.R
\name{findPathwayUrlsByOrcid}
\alias{findPathwayUrlsByOrcid}
\title{Find Pathway URLs By ORCID}
\usage{
findPathwayUrlsByOrcid(query = NULL)
}
\arguments{
\item{query}{The \code{character} ORCID to search for.}
}
\value{
A \code{list} of lists
}
\description{
Retrieve list of pathway URLs containing the query ORCID
}
\examples{
{
findPathwayUrlsByOrcid(' 0000-0001-9773-4008')
}
}
\seealso{
findPathwaysByOrcid
}
