% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{getCounts}
\alias{getCounts}
\title{Get Counts for WikiPathways Stats}
\usage{
getCounts()
}
\value{
A \code{data.frame} of counts
}
\description{
Retrieve information about various total counts at WikiPathways.
}
\examples{
{
getCounts()
}
}
