% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurationStatus.R
\name{getCurationStatus}
\alias{getCurationStatus}
\title{Get Curation Status of a Pathway}
\usage{
getCurationStatus(pathway)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway, e.g. WP554}
}
\value{
A \code{data.frame} of status details
}
\description{
Retrieve information about curation status for a specific pathway.
}
\examples{
{
getCurationStatus('WP554')
}
}
