% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurationTags.R
\name{getCurationTagNames}
\alias{getCurationTagNames}
\title{Get Curation Tag Names on a Pathway}
\usage{
getCurationTagNames(pathway)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway to download, e.g. WP4}
}
\value{
List of tag names
}
\description{
Retrieve names of curation tags for a specific pathway.
}
\examples{
{
getCurationTagNames('WP554')
}
}
