% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurationTags.R
\name{getCurationTags}
\alias{getCurationTags}
\title{Get Curation Tags on a Pathway}
\usage{
getCurationTags(pathway)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway to download, e.g. WP4}
}
\value{
A \code{list} of tag name, display name, revision, text, timestampe and user
}
\description{
Retrieve information about curation tags for a specific pathway.
}
\examples{
{
getCurationTags('WP554')
}
}
