% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurationTags.R
\name{getEveryCurationTag}
\alias{getEveryCurationTag}
\title{Get Every Instance of a Curation Tag}
\usage{
getEveryCurationTag(tag)
}
\arguments{
\item{tag}{Official name of curation tag, e.g., "Curation:FeaturedPathway"}
}
\value{
List of tag name, display name, revision, text, timestampe and user
}
\description{
Retrieve information about every instance of a given curation tag.
}
\examples{
\donttest{
getEveryCurationTag('Curation:FeaturedPathway')
}
}
