% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntologyTerms.R
\name{getOntologyTermIds}
\alias{getOntologyTermIds}
\title{Get Ontology Term IDs by Pathway}
\usage{
getOntologyTermIds(pathway)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway to download, e.g. WP4}
}
\value{
List of term identifiers
}
\description{
Retrieve identifiers of ontology terms for a specific pathway.
}
\examples{
{
getOntologyTermIds('WP554')
}
}
