% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntologyTerms.R
\name{getOntologyTermNames}
\alias{getOntologyTermNames}
\title{Get Ontology Term Names by Pathway}
\usage{
getOntologyTermNames(pathway)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway to download, e.g. WP4}
}
\value{
List of term names
}
\description{
Retrieve names of ontology terms for a specific pathway.
}
\examples{
{
getOntologyTermNames('WP554')
}
}
