% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntologyTerms.R
\name{getOntologyTerms}
\alias{getOntologyTerms}
\title{Get Ontology Terms by Pathway}
\usage{
getOntologyTerms(pathway = NULL)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway, e.g. WP554. If
NULL, then ontology term information for all pathways is returned.}
}
\value{
A \code{data.frame} pathway id and term information
}
\description{
Retrieve information about ontology terms for a specific pathway.
}
\examples{
{
getOntologyTerms('WP554')
}
}
