% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathwayHistory.R
\name{getPathwayHistory}
\alias{getPathwayHistory}
\title{Get Pathway History}
\usage{
getPathwayHistory(pathway = NULL, timestamp = NULL)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway, e.g. WP554}

\item{timestamp}{<ignored>}
}
\value{
Opens the GitHub history for a pathway
}
\description{
View the revision history of a pathway.
}
\examples{
{
getPathwayHistory('WP554')
}
}
