% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCommunities.R
\name{getPathwayIdsByCommunity}
\alias{getPathwayIdsByCommunity}
\title{Get Pathway IDs By Community}
\usage{
getPathwayIdsByCommunity(community_tag = NULL)
}
\arguments{
\item{community_tag}{Abbreviated name of community}
}
\value{
A \code{list} of pathway IDs
}
\description{
Retrieve the list of pathway IDs per community
}
\examples{
{
getPathwayIdsByCommunity("AOP")
}

}
