% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurationStatus.R
\name{getPathwayIdsByCurationTag}
\alias{getPathwayIdsByCurationTag}
\title{DEPRECATED: Get Pathway WPIDs by Curation Tag}
\usage{
getPathwayIdsByCurationTag(tag)
}
\arguments{
\item{tag}{(\code{character}) Official name of curation tag, e.g., "Curation:FeaturedPathway"}
}
\value{
A \code{list} of pathway WPIDs
}
\description{
This function is provided for compatibility with older 
web services only and will be defunct at the next release.
}
