% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntologyTerms.R
\name{getPathwayIdsByParentOntologyTerm}
\alias{getPathwayIdsByParentOntologyTerm}
\title{Get Pathway WPIDs by Parent Ontology Term}
\usage{
getPathwayIdsByParentOntologyTerm(term = NULL)
}
\arguments{
\item{term}{(\code{character}) Official name of ontology term, e.g., "signaling pathway"}
}
\value{
A \code{list} of pathway WPIDs
}
\description{
Retrieve pathway WPIDs for every pathway with a child term of given ontology term
}
\examples{
{
getPathwayIdsByParentOntologyTerm('signaling pathway')
}
}
