% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathwayInfo.R
\name{getPathwayInfo}
\alias{getPathwayInfo}
\title{Get Pathway Info}
\usage{
getPathwayInfo(pathway = NULL)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway to download, 
e.g. WP554. If NULL, then all pathways are returned.}
}
\value{
A \code{dataframe} of pathway WPID, URL, name, species, revision,
authors, description, and citedIn
}
\description{
Retrieve information for a specific pathway
}
\examples{
{
getPathwayInfo('WP554')
}
}
