% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCommunities.R
\name{getPathwayNamesByCommunity}
\alias{getPathwayNamesByCommunity}
\title{Get Pathway Names By Community}
\usage{
getPathwayNamesByCommunity(community_tag = NULL)
}
\arguments{
\item{community_tag}{Abbreviated name of community}
}
\value{
A \code{list} of pathway names
}
\description{
Retrieve the list of pathway names per community
}
\examples{
{
getPathwayNamesByCommunity("AOP")
}

}
