% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCommunities.R
\name{getPathwaysByCommunity}
\alias{getPathwaysByCommunity}
\title{Get Pathways By Community}
\usage{
getPathwaysByCommunity(community_tag = NULL)
}
\arguments{
\item{community_tag}{Abbreviated name of community}
}
\value{
A \code{data.frame} of pathway information
}
\description{
Retrieve pathways per community
}
\examples{
{
getPathwaysByCommunity("AOP")
}

}
