% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurationTags.R
\name{getPathwaysByCurationTag}
\alias{getPathwaysByCurationTag}
\title{Get Pathways by Curation Tag}
\usage{
getPathwaysByCurationTag(tag)
}
\arguments{
\item{tag}{(\code{character}) Official name of curation tag, e.g., "Curation:FeaturedPathway"}
}
\value{
A \code{list} of pathway information, including WPID, url, name, species and revision
}
\description{
Retrieve pathway information for every pathway with a given curation tag.
}
\examples{
{
getPathwaysByCurationTag('Curation:FeaturedPathway')
}
}
