% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntologyTerms.R
\name{getPathwaysByOntologyTerm}
\alias{getPathwaysByOntologyTerm}
\title{Get Pathways by Ontology Term}
\usage{
getPathwaysByOntologyTerm(term = NULL)
}
\arguments{
\item{term}{(\code{character}) Official ID of ontology term, e.g., "PW:0000045"}
}
\value{
A \code{data.frame} of pathway information
}
\description{
Retrieve pathway information for every pathway with a given ontology term.
}
\examples{
{
getPathwaysByOntologyTerm('PW:0000045')
}
}
