% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntologyTerms.R
\name{getPathwaysByParentOntologyTerm}
\alias{getPathwaysByParentOntologyTerm}
\title{Get Pathways by Parent Ontology Term}
\usage{
getPathwaysByParentOntologyTerm(term = NULL)
}
\arguments{
\item{term}{(\code{character}) Official name of ontology term, e.g., "signaling pathway"}
}
\value{
A \code{data.frame} of pathway information
}
\description{
Retrieve pathway information for every pathway with a child term of given ontology term.
}
\examples{
{
getPathwaysByParentOntologyTerm('signaling pathway')
}
}
