% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCommunities.R
\name{getPathwayUrlsByCommunity}
\alias{getPathwayUrlsByCommunity}
\title{Get Pathway URLs By Community}
\usage{
getPathwayUrlsByCommunity(community_tag = NULL)
}
\arguments{
\item{community_tag}{Abbreviated name of community}
}
\value{
A \code{list} of pathway URLs
}
\description{
Retrieve the list of pathway URLs per community
}
\examples{
{
getPathwayUrlsByCommunity("AOP")
}

}
