% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecentChanges.R
\name{getRecentChangesIds}
\alias{getRecentChangesIds}
\title{Get WPIDs of Recent Changes}
\usage{
getRecentChangesIds(timestamp)
}
\arguments{
\item{timestamp}{(8 digits, YYYYMMDD) Limit by time, only pathways changed 
after the given date, e.g., 20180201 for changes since Feb 1st, 2018.}
}
\value{
A \code{list} of WPIDs
}
\description{
Retrieve WPIDs of recently changed pathways at WikiPathways.
}
\examples{
{
getRecentChangesIds('20180201')
}
}
