% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecentChanges.R
\name{getRecentChangesNames}
\alias{getRecentChangesNames}
\title{Get Pathway Names of Recent Changes}
\usage{
getRecentChangesNames(timestamp)
}
\arguments{
\item{timestamp}{(8 digits, YYYYMMDD) Limit by time, only pathways changed 
after the given date, e.g., 20180201 for changes since Feb 1st, 2018.}
}
\value{
A \code{list} of pathway names. Note: pathway deletions will be listed as blank names.
}
\description{
Retrieve names of recently changed pathways at WikiPathways.
}
\examples{
{
getRecentChangesNames('20180201')
}
}
