% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCommunities.R
\name{listCommunities}
\alias{listCommunities}
\title{List Communities}
\usage{
listCommunities()
}
\value{
A \code{data.frame} of community information
}
\description{
Retrieve the list of communities hosted by WikiPathways
}
\examples{
{
listCommunities()
}

}
