\name{listOrganisms}
\alias{listOrganisms}
\title{
Lists the organisms available from the WikiPathways API server
}
\description{
Lists the organisms available from the WikiPathways API server
}
\usage{
listOrganisms()
}
\value{
 \item{data}{the available organisms}
}
\examples{
organisms <- listOrganisms()
}
\author{
Egon Willighagen
}
