% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPathways.R
\name{listPathwayIds}
\alias{listPathwayIds}
\title{List Pathway WPIDs}
\usage{
listPathwayIds(organism = "")
}
\arguments{
\item{organism}{(\code{character}, optional) A particular species. 
See \link{listOrganisms}. 
Default is all species.}
}
\value{
A \code{list} of WPIDs
}
\description{
Retrieve list of pathway WPIDs per species.
}
\details{
Basically returns a subset of \link{listPathways} result
}
\examples{
{
listPathwayIds('Mus musculus')
}
}
\seealso{
listPathways
}
