% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPathways.R
\name{listPathways}
\alias{listPathways}
\title{List Pathways}
\usage{
listPathways(organism = "")
}
\arguments{
\item{organism}{(optional) A particular species. See \link{listOrganisms}. 
Default is all species.}
}
\value{
list of lists
}
\description{
Retrieve list of pathways per species, including WPID, name, species,
URL and latest revision number.
}
\examples{
{
listPathways('Mus musculus')
}
}
