% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPathways.R
\name{listPathwayUrls}
\alias{listPathwayUrls}
\title{List Pathway URLs}
\usage{
listPathwayUrls(organism = "")
}
\arguments{
\item{organism}{(optional) A particular species. See \link{listOrganisms}. 
Default is all species.}
}
\value{
list of URLs
}
\description{
Retrieve list of pathway URLs per species.
}
\details{
Basically returns a subset of \link{listPathways} result
}
\examples{
\donttest{
listPathwayUrls('Mus musculus')
}
}
