% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMT.R
\name{readGMT}
\alias{readGMT}
\title{Read GMT File}
\usage{
readGMT(file)
}
\arguments{
\item{file}{Path to GMT file}
}
\value{
Data frame of term-gene associations
}
\description{
Reads any generic GMT file to produce a data frame of 
term-gene associations useful in enrichment analyses and other 
applications.
}
\details{
The returned data frame includes only terms and genes. If you want
another data frame with terms and names, then see readGMTnames.
}
\examples{
\donttest{
readGMT(system.file("extdata","my_gmt_file.gmt", package="rWikiPathways"))
}
}
\seealso{
readGMTnames
}
