% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMT.R
\name{readGMTnames}
\alias{readGMTnames}
\title{Read GMT File for Names}
\usage{
readGMTnames(file)
}
\arguments{
\item{file}{Path to GMT file}
}
\value{
Data frame of term-namee associations
}
\description{
Reads any generic GMT file to produce a data frame of 
term-name associations useful in enrichment analyses and other 
applications.
}
\details{
The returned data frame includes only terms and names. If you want
another data frame with terms and genes, then see readGMT.
}
\examples{
\donttest{
#readGMTnames(system.file("extdata","my_gmt_file.gmt", package="rWikiPathways"))
#readGMTnames("path_to_your_gmt_file.gmt")
}
}
\seealso{
readGMT
}
