% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wikipathwaysAPI}
\alias{wikipathwaysAPI}
\title{Open Swagger docs for WikiPathways API}
\usage{
wikipathwaysAPI(base.url = .baseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain.}
}
\value{
New tab in default browser
}
\description{
Opens swagger docs in default browser for available API calls.
}
\examples{
{
wikipathwaysAPI()
}
}
