% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wikipathwaysGET}
\alias{wikipathwaysGET}
\title{WikiPathways GET}
\usage{
wikipathwaysGET(
  operation,
  parameters = NULL,
  format = c("json", "xml", "html", "jpg", "pdf", "dump"),
  base.url = .baseUrl
)
}
\arguments{
\item{operation}{A \code{character} string to be converted to the query 
namespace}

\item{parameters}{A named \code{list} of values to be converted to query 
parameters}

\item{format}{(\code{character}) The format of the return, e.g., json 
(default), xml, html, jpg, pdf, dump}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain.}
}
\value{
query result content
}
\description{
Constructs the query, makes GET call and processes the result
}
\examples{
{
wikipathwaysGET('listOrganisms')
}
}
