% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPathwayArchive.R
\name{downloadPathwayArchive}
\alias{downloadPathwayArchive}
\title{Download Pathway Archive}
\usage{
downloadPathwayArchive(date = "current", format = "gpml")
}
\arguments{
\item{date}{The timestamp for a monthly release or "current" (default) for 
latest release.}

\item{format}{Either gpml (default), gmt, index, rdf or svg.}
}
\value{
Open tab in default browser
}
\description{
Access the monthly archives of pathway content from WikiPathways.
}
\examples{
\donttest{
downloadPathwayArchive()
downloadPathwayArchive(format="gmt")
downloadPathwayArchive("20171010","svg")
}
}
