% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwayByText.R
\name{findPathwayIdsByText}
\alias{findPathwayIdsByText}
\title{Find Pathway WPIDs By Text}
\usage{
findPathwayIdsByText(query)
}
\arguments{
\item{query}{A \code{character} string to search for, e.g., "cancer"}
}
\value{
A \code{list} of WPIDs
}
\description{
Retrieve list of pathway WPIDs containing the query text.
}
\examples{
{
findPathwayIdsByText('cancer')
}
}
\seealso{
findPathwaysByText
}
