% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwayByText.R
\name{findPathwayIdsByText}
\alias{findPathwayIdsByText}
\title{Find Pathway WPIDs By Text}
\usage{
findPathwayIdsByText(query)
}
\arguments{
\item{query}{The string to search for, e.g., "cancer"}
}
\value{
List of WPIDs
}
\description{
Retrieve list of pathway WPIDs containing the query text.
}
\examples{
{
findPathwayIdsByText('cancer')
}
}
