% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwayByText.R
\name{findPathwayNamesByText}
\alias{findPathwayNamesByText}
\title{Find Pathway Names By Text}
\usage{
findPathwayNamesByText(query)
}
\arguments{
\item{query}{The string to search for, e.g., "cancer"}
}
\value{
List of lists
}
\description{
Retrieve list of pathway names containing the query text.
}
\details{
Note: there will be multiple listings of the same pathway name if 
copies exist for multiple species.
}
\examples{
{
findPathwayNamesByText('cancer')
}
}
