% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwaysByLiterature.R
\name{findPathwaysByLiterature}
\alias{findPathwaysByLiterature}
\title{Find Pathways By Literature}
\usage{
findPathwaysByLiterature(query = NULL)
}
\arguments{
\item{query}{The \code{character} string to search for, e.g., a PMID, title 
keyword, journal abbreviation, year, or author name.}
}
\value{
A \code{dataframe} of pathway attributes including the matching
citations
}
\description{
Retrieve pathways containing the query citation.
}
\examples{
{
findPathwaysByLiterature('15134803')
findPathwaysByLiterature('Schwartz GL')
findPathwaysByLiterature('Eur J Pharmacol')
findPathwaysByLiterature('antihypertensive drug responses')
}
}
