% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwayByText.R
\name{findPathwaysByText}
\alias{findPathwaysByText}
\title{Find Pathways By Text}
\usage{
findPathwaysByText(query)
}
\arguments{
\item{query}{A \code{character} string to search for, e.g., "cancer"}
}
\value{
A \code{dataframe} of pathway attributes in addition to query result score
}
\description{
Retrieve pathways containing the query text.
}
\details{
The score is from a lucene index search engine, ranging from 0 to 
1 with higher scores for better matches.
}
\examples{
{
findPathwaysByText('cancer')
}
}
