% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwayByText.R
\name{findPathwaysByText}
\alias{findPathwaysByText}
\title{Find Pathways By Text}
\usage{
findPathwaysByText(query)
}
\arguments{
\item{query}{The string to search for, e.g., "cancer"}
}
\value{
List of lists
}
\description{
Retrieve a list of pathways containing the query text.
}
\examples{
\donttest{
findPathwaysByText('cancer')
}
}
