% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwaysByLiterature.R
\name{findPathwayUrlsByLiterature}
\alias{findPathwayUrlsByLiterature}
\title{Find Pathway URLs By Literature}
\usage{
findPathwayUrlsByLiterature(query)
}
\arguments{
\item{query}{The \code{character} string to search for, e.g., a PMID, title keyword or author name.}
}
\value{
A \code{list} of lists
}
\description{
Retrieve list of pathway URLs containing the query citation.
}
\examples{
{
findPathwayUrlsByLiterature('19649250')
findPathwayUrlsByLiterature('smith')
findPathwayUrlsByLiterature('cancer')
}
}
