% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwaysByText.R
\name{findPathwayUrlsByText}
\alias{findPathwayUrlsByText}
\title{Find Pathway URLs By Text}
\usage{
findPathwayUrlsByText(query = NULL, field = NULL)
}
\arguments{
\item{query}{A \code{character} string to search for, e.g., "cancer"}

\item{field}{Optional \code{character} string to restrict search to a single
field, e.g., title, description, wpid, species, revision, authors,
datanodes, annotations, or citedIn.}
}
\value{
A \code{list} of urls
}
\description{
Retrieve list of pathway URLs containing the query text.
}
\examples{
{
findPathwayUrlsByText('cancer')
}
}
\seealso{
findPathwaysByText
}
