% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathwayByText.R
\name{findPathwayUrlsByText}
\alias{findPathwayUrlsByText}
\title{Find Pathway URLs By Text}
\usage{
findPathwayUrlsByText(query)
}
\arguments{
\item{query}{The string to search for, e.g., "cancer"}
}
\value{
List of lists
}
\description{
Retrieve list of pathway URLs containing the query text.
}
\examples{
\donttest{
findPathwayUrlsByText('cancer')
}
}
