% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColoredPathway.R
\name{getColoredPathway}
\alias{getColoredPathway}
\title{Defunct: Get Colored Pathway}
\usage{
getColoredPathway(
  pathway,
  revision = 0,
  graphId = NULL,
  color = NULL,
  fileType = c("svg", "png", "pdf")
)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway to download, e.g. WP4}

\item{revision}{(\code{integer}, optional) Number to indicate a specific revision to download}

\item{graphId}{A \code{character} string or \code{vector} indicating the nodes to color}

\item{color}{(optional) String or vector indicating the highlighting color, e.g., #FF8855.
Default is red. You can provide a single color for mutiple nodes; otherwise
color list and graphId must be the same length.}

\item{fileType}{(optional) Image file format, e.g., svg (default), png or pdf.}
}
\value{
Image file
}
\description{
This service is no longer available and is defunct in 
this release and will be removed in the next release.
}
