% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColoredPathway.R
\name{getColoredPathway}
\alias{getColoredPathway}
\title{Get Colored Pathway}
\usage{
getColoredPathway(pathway, revision = 0, graphId, color = NULL,
  fileType = "svg")
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway to download, e.g. WP4}

\item{revision}{(optional) Number to indicate a specific revision to download}

\item{graphId}{String or vector indicating the nodes to color (only single values are supported at this time)}

\item{color}{(optional) String or vector indicating the highlighting color, e.g., #FF8855.
Default is red. You can provide a single color for mutiple nodes; otherwise
color list and graphId must be the same length.}

\item{fileType}{(optional) Image file format (only 'svg' supported at this time)}
}
\value{
Image file
}
\description{
Retrieve a pathway image file with specified nodes colored by specified colors
}
\examples{
\donttest{
  svg = getColoredPathway(pathway="WP554", graphId="ef1f3");
  svg = getColoredPathway(pathway="WP554", graphId=c("ef1f3","e68e0"));
  svg = getColoredPathway(pathway="WP554", graphId=c("ef1f3","e68e0"),
                          color="00FF00");
  svg = getColoredPathway(pathway="WP554", graphId=c("ef1f3","e68e0"),
                          color=c("FF0000", "0000FF"));
  writeLines(svg, "pathway.svg")
}
}
