% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurationStatus.R
\name{getCurationTagNames}
\alias{getCurationTagNames}
\title{DEPRECATED: Get Curation Tag Names on a Pathway}
\usage{
getCurationTagNames(pathway)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway to download, e.g. WP4}
}
\value{
A \code{list} of tag names
}
\description{
This function is provided for compatibility with older 
web services only and will be defunct at the next release.
}
\examples{
{
getCurationTagNames('WP554')
}
}
