% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurationStatus.R
\name{getCurationTags}
\alias{getCurationTags}
\title{DEPRECATED: Get Curation Tags on a Pathway}
\usage{
getCurationTags(pathway)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway to download, e.g. WP4}
}
\value{
A \code{list} of tag name, display name, revision, text, timestampe and user
}
\description{
This function is provided for compatibility with older 
web services only and will be defunct at the next release.
}
\examples{
{
getCurationTags('WP554')
}
}
