% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurationStatus.R
\name{getEveryCurationTag}
\alias{getEveryCurationTag}
\title{DEPRECATED: Get Every Instance of a Curation Tag}
\usage{
getEveryCurationTag(tag)
}
\arguments{
\item{tag}{(\code{character}) Official name of curation tag, e.g., "Curation:FeaturedPathway"}
}
\value{
A \code{list} of tag name, display name, revision, text, timestampe and user
}
\description{
This function is provided for compatibility with older 
web services only and will be defunct at the next release.
}
