% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntologyTerms.R
\name{getOntologyTerms}
\alias{getOntologyTerms}
\title{Get Ontology Terms by Pathway}
\usage{
getOntologyTerms(pathway)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway to download, e.g. WP4}
}
\value{
A \code{list} of tag name, display name, revision, text, timestampe and user
}
\description{
Retrieve information about ontology terms for a specific pathway.
}
\examples{
{
getOntologyTerms('WP554')
}
}
