% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathwayHistory.R
\name{getPathwayHistory}
\alias{getPathwayHistory}
\title{Get Pathway History}
\usage{
getPathwayHistory(pathway, timestamp)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway to download, e.g. WP4}

\item{timestamp}{Limit by time, only history items after the given time, 
e.g., 20180201 for revisions since Feb 1st, 2018. Any length of timestamp is 
accepted, upto 14 digits, e.g., 2018, 201802, 20180201063011, etc.}
}
\value{
List of revisions, including user and comment
}
\description{
Retrieve the revision history of a pathway.
}
\examples{
\donttest{
getPathwayHistory('WP554')
}
}
