% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurationTags.R
\name{getPathwayIdsByCurationTag}
\alias{getPathwayIdsByCurationTag}
\title{Get Pathway WPIDs by Curation Tag}
\usage{
getPathwayIdsByCurationTag(tag)
}
\arguments{
\item{tag}{Official name of curation tag, e.g., "Curation:FeaturedPathway"}
}
\value{
List of pathway WPIDs
}
\description{
Retrieve pathway WPIDs for every pathway with a given curation tag.
}
\examples{
\donttest{
getPathwayIdsByCurationTag('Curation:FeaturedPathway')
}
}
