% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntologyTerms.R
\name{getPathwayIdsByOntologyTerm}
\alias{getPathwayIdsByOntologyTerm}
\title{Get Pathway WPIDs by Ontology Term}
\usage{
getPathwayIdsByOntologyTerm(term)
}
\arguments{
\item{term}{Official name of ontology term, e.g., "PW:0000045"}
}
\value{
List of pathway WPIDs
}
\description{
Retrieve pathway WPIDs for every pathway with a given ontology term.
}
\examples{
{
getPathwayIdsByOntologyTerm('PW:0000045')
}
}
