% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathwayInfo.R
\name{getPathwayInfo}
\alias{getPathwayInfo}
\title{Get Pathway Info}
\usage{
getPathwayInfo(pathway)
}
\arguments{
\item{pathway}{WikiPathways identifier (WPID) for the pathway to download, e.g. WP4}
}
\value{
A \code{dataframe} of pathway WPID, URL, name, species and latest revision
}
\description{
Retrieve information for a specific pathway
}
\examples{
{
getPathwayInfo('WP554')
}
}
