% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurationStatus.R
\name{getPathwaysByCurationTag}
\alias{getPathwaysByCurationTag}
\title{DEPRECATED: Get Pathways by Curation Tag}
\usage{
getPathwaysByCurationTag(tag)
}
\arguments{
\item{tag}{(\code{character}) Official name of curation tag, e.g., "Curation:FeaturedPathway"}
}
\value{
A \code{list} of pathway information, including WPID, url, name, species and revision
}
\description{
This function is provided for compatibility with older 
web services only and will be defunct at the next release.
}
\examples{
{
getPathwaysByCurationTag('Curation:FeaturedPathway')
}
}
