% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntologyTerms.R
\name{getPathwaysByOntologyTerm}
\alias{getPathwaysByOntologyTerm}
\title{Get Pathways by Ontology Term}
\usage{
getPathwaysByOntologyTerm(term)
}
\arguments{
\item{term}{(\code{character}) Official name of ontology term, e.g., "PW:0000045"}
}
\value{
A \code{list} of pathway information, including WPID, url, name, species and revision
}
\description{
Retrieve pathway information for every pathway with a given ontology term.
}
\examples{
{
getPathwaysByOntologyTerm('PW:0000045')
}
}
