% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecentChanges.R
\name{getRecentChangesIds}
\alias{getRecentChangesIds}
\title{Get WPIDs of Recent Changes}
\usage{
getRecentChangesIds(timestamp)
}
\arguments{
\item{timestamp}{(14 digits, YYYYMMDDhhmmss) Limit by time, only history items 
after the given time, e.g., 20180201000000 for changes since Feb 1st, 2018.}
}
\value{
List of WPIDs
}
\description{
Retrieve WPIDs of recently changed pathways at WikiPathways.
}
\examples{
{
getRecentChangesIds('20180201000000')
}
}
