% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecentChanges.R
\name{getRecentChangesNames}
\alias{getRecentChangesNames}
\title{Get Pathway Names of Recent Changes}
\usage{
getRecentChangesNames(timestamp)
}
\arguments{
\item{timestamp}{(14 digits, YYYYMMDDhhmmss) Limit by time, only history items 
after the given time, e.g., 20180201000000 for changes since Feb 1st, 2018.}
}
\value{
List of pathway names. Note: pathway deletions will be listed as blank names.
}
\description{
Retrieve names of recently changed pathways at WikiPathways.
}
\examples{
\donttest{
getRecentChangesNames('20180201000000')
}
}
