% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOrganisms.R
\name{listOrganisms}
\alias{listOrganisms}
\title{List Organisms}
\usage{
listOrganisms()
}
\value{
list of organisms
}
\description{
Retrieve the list of organisms supported by WikiPathways
}
\examples{
\donttest{
listOrganisms()
}
}
