% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wikipathwaysAPI}
\alias{wikipathwaysAPI}
\title{DEPRECATED: Open Swagger docs for WikiPathways API}
\usage{
wikipathwaysAPI(base.url = .baseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain.}
}
\value{
New tab in default browser
}
\description{
This function is provided for compatibility with older 
web services only and will be defunct at the next release.
}
\examples{
{
wikipathwaysAPI()
}
}
