% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wikipathwaysGET}
\alias{wikipathwaysGET}
\title{WikiPathways GET}
\usage{
wikipathwaysGET(operation, parameters = NULL, format = "json",
  base.url = .baseUrl)
}
\arguments{
\item{operation}{A string to be converted to the query namespace}

\item{parameters}{A named list of values to be converted to query parameters}

\item{format}{The format of the return, e.g., json (default), xml, html, jpg, pdf, dump}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain.}
}
\value{
query result content
}
\description{
Constructs the query, makes GET call and processes the result
}
\examples{
{
wikipathwaysGET('listOrganisms')
}
}
